backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless
scikit-learn
reportlab
numpy
torch
torchvision
openai
faiss-cpu
sentence-transformers
python-magic
aiofiles
